requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
pure-python-adb>=0.3.0
scipy>=1.11.0
pandas>=2.0.0
pygame>=2.5.0
loguru>=0.7.0
pyyaml>=6.0.0
tqdm>=4.65.0 